requirements.txt
aerich==0.7.2
aiodns==3.0.0
aiofiles==23.2.1
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.17.0
amqp==5.2.0
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
arq==0.25.0
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.28.0
attrs==23.1.0
Babel==2.13.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.2.0
Brotli==1.1.0
bs4==0.0.1
celery==5.3.5
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
decorator==5.1.1
dictdiffer==0.9.0
dnspython==2.4.2
email-validator==2.1.0.post1
eralchemy2==1.3.8
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.103.1
fastapi-admin==1.0.4
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
hiredis==2.2.3
httptools==0.6.0
idna==3.4
ipython==8.17.2
iso8601==1.1.0
jedi==0.19.0
Jinja2==3.1.2
kombu==5.3.3
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
multidict==6.0.4
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.2
parso==0.8.3
pendulum==2.1.2
pexpect==4.8.0
phonenumbers==8.13.24
playwright==1.37.0
prompt-toolkit==3.0.39
psycopg==3.1.10
psycopg-binary==3.1.10
psycopg-pool==3.1.7
psycopg2-binary==2.9.7
ptpython==3.0.23
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.0
pycares==4.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyee==9.0.4
Pygments==2.16.1
pygraphviz==1.11
PyNaCl==1.5.0
pypika-tortoise==0.1.6
pypostalcode==0.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
rapidfuzz==3.3.1
redis==5.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
termcolor==2.3.0
thefuzz==0.20.0
toml==0.10.2
tomlkit==0.12.1
tortoise-cli==0.1.2
tortoise-orm==0.20.0
traitlets==5.13.0
typing_extensions==4.7.1
tzdata==2023.3
ua-generator==0.1.8
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
validators==0.22.0
vine==5.1.0
watchfiles==0.20.0
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
